+++ nodes/EasyRedmine/operations/CreateOperation.ts
@@ -12,6 +12,7 @@ import {
 	TimeEntryCreateOptions,
 	UserCreateOptions,
 } from './CreateModel';
+import { convertToEasyDate } from '../utils/ConvertToEasyDate';
 
 function convertCustomFields(options: CreateOptionsWithCustomFields): CustomField[] | undefined {
 	return options.customFields?.field.map((customField) => ({
@@ -80,8 +81,8 @@ function createBodyForIssue(this: IExecuteFunctions, itemIndex: number): { [key:
 			assigned_to_id: options.assignedToId,
 			estimated_hours: options.estimatedHours,
 			done_ratio: options.doneRatio,
-			start_date: options.startDate,
-			due_date: options.dueDate,
+			start_date: convertToEasyDate(options.startDate),
+			due_date: convertToEasyDate(options.dueDate),
 			is_private: options.isPrivate,
 			priority_id: options.priorityId,
 			status_id: options.statusId,